fix too many clients error
