+++ backend/main.py
@@ -7,10 +7,8 @@ from fastapi.middleware.cors import CORSMiddleware
 from fastapi.middleware.gzip import GZipMiddleware
 
 from api import routes
-from services.scripts.simulate_vitals import (manage_simulations,
-                                              simulate_patient)
 from services.simulation_manager import SimulationManager
-from src.db.models.db_connection import TimescaleDBClient
+from src.db.db_connection import TimescaleDBClient
 from src.db.repository.patients_repository import PatientRepository
 from src.db.repository.vital_records_repository import VitalsRepository
 from src.db.scripts.populate_patients import populate_patients_if_empty
