+++ backend/services/simulation_manager.py
@@ -4,7 +4,7 @@ from typing import Dict, Set
 
 from services.scripts.simulate_vitals import (VitalRecordSimulator,
                                               simulate_patient)
-from src.db.models.db_connection import TimescaleDBClient
+from src.db.db_connection import TimescaleDBClient
 from src.db.repository.vital_records_repository import VitalsRepository
