+++ backend/src/db/repository/patients_repository.py
@@ -1,7 +1,7 @@
 from datetime import datetime
 from typing import List, Optional
 
-from src.db.models.db_connection import TimescaleDBClient
+from src.db.db_connection import TimescaleDBClient
 from src.db.models.patient import Patient
